fix: resolve change request and refactor the code
