+++ client/src/components/createMatchStyled/index.ts
@@ -1,0 +1,6 @@
+export {
+  CreateMatchButtons,
+  DialogInputsBox,
+  StyledAutocomplete,
+  CreateMatchImg,
+} from './createMatchStyled';
