+++ client/src/components/styledRootComponent/index.ts
@@ -1,0 +1,1 @@
+export { StyledNavBox, StyledSearchInput } from './Nav';
